+++ cjwpandasmodule/i18n/__init__.py
@@ -1,3 +1,0 @@
-from .funcs import _trans_cjwpandasmodule
-
-__all__ = ["_trans_cjwpandasmodule"]